requirements.txt
streamlit
streamlit-drawable-canvas
diffusers
transformers
accelerate
safetensors
torch
Pillow
numpy
opencv-python
